Generative AI/Ecom_Chatbot/requirements.txt
langchain-astradb
langchain 
datasets
pypdf
python-dotenv
flask
langchain-google-genai

-e .